requirements.txt
astroid==1.6.5
Django==1.9.2
djangorestframework==3.3.2
djangorestframework-jwt==1.8.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.6.1
pyjade==4.0.0
PyJWT==1.4.0
pylint==1.9.2
six==1.11.0
wrapt==1.10.11
whitenoise==3.3.1
